refactor(sentry): extract reverse-family verification from queue-io.mjs
scripts/sentry-autofix-queue-io.mjs sat at 621 lines, over the recurring-review
600-line soft cap (Codex flagged it — the #1784 extraction moved the violation
into this file). Move the reverse `in:comments` verification leg
(`reverseVerifyFamilies` + its `readVerdictCached` helper and the
MAX_REVERSE_PROBE_QUERIES / REVERSE_SEARCH_LIMIT / MAX_REVERSE_VERIFY_READS
budgets) into a new scripts/sentry-autofix-reverse-verify.mjs. queue-io keeps the
forward queue reads (listCodeFixStubs / listHandledShortIds / readStub /
openAutofixPrExists / isOwnHeadPr) and is now 392 lines.

Behavior unchanged. The reverse leg's tests stay in the select suite (no new
scripts/sentry-*.test.mjs, so no suite-registration cascade); the select suite
drives the reverse probe through mocked runGh and passes 65/65.
SENTRY_TRIAGE_QUEUE_LABEL is now exported from queue-io so the new module narrows
its probe the same way; sentry-autofix-family-resolve.mjs and the select suite
import reverseVerifyFamilies + the reverse budgets from the new module, and the
quality gate routes the new module to the select suite.
