+++ scripts/sentry-autofix-queue-io.mjs
@@ -14,11 +14,7 @@
 
 import { spawn } from "node:child_process";
 
-import {
-  isValidShortId,
-  parseShortId,
-  resolveVerdict,
-} from "./sentry-triage-project-core.mjs";
+import { parseShortId } from "./sentry-triage-project-core.mjs";
 import {
   ARCHIVED_LABEL,
   CODE_FIX_VERDICT_LABEL,
@@ -27,16 +23,15 @@ import {
   PROJECTED_LABEL,
 } from "./sentry-triage-ingest.mjs";
 import { autofixBranchName } from "./sentry-autofix-finalize.mjs";
-import {
-  declaredFamilyIds,
-  familyKey,
-  isLocalFamilyId,
-} from "./sentry-autofix-family.mjs";
+import { familyKey } from "./sentry-autofix-family.mjs";
 
 // The queue-membership label every triage stub carries (the queue contract's
 // `LABEL_DEFINITIONS` self-heals it). Both per-id lookups below narrow their
 // search to it so a random issue that merely quotes a SHORT-ID cannot match.
-const SENTRY_TRIAGE_QUEUE_LABEL = "sentry-triage";
+// Exported so the extracted reverse-verify leg
+// (sentry-autofix-reverse-verify.mjs) narrows its `in:comments` probe the same
+// way.
+export const SENTRY_TRIAGE_QUEUE_LABEL = "sentry-triage";
 
 // The verdict label the select scans for — re-exported from the ingest's single
 // source of truth (`code-fix` maps to this label), so a future rename can't
@@ -290,230 +285,6 @@ export async function listHandledShortIds(
     }
   }
   return [...handled];
-}
-
-/**
- * Read a stub and resolve its verdict through the SAME authorship/regression
- * fence the label step uses, memoised by issue number for the run. Returns the
- * parsed verdict, or `null` when the stub cannot be read or carries no usable
- * fenced verdict (a comment mention that is not a real verdict). Never throws.
- */
-async function readVerdictCached(runGh, repo, number, cache) {
-  const key = String(number);
-  if (cache.has(key)) return cache.get(key);
-  let parsed;
-  try {
-    const full = await readStub(runGh, repo, number);
-    parsed = resolveVerdict(full, number).parsed;
-  } catch {
-    parsed = null;
-  }
-  cache.set(key, parsed);
-  return parsed;
-}
-
-/**
- * How many distinct reverse `in:comments` probes one RUN may issue — the bug-B
- * mirror of MAX_HANDLED_ID_QUERIES, and the missing sibling of it before PR
- * #1810's follow-up. `probeIds` is the union of the finalists' family members,
- * and a family can hold up to MAX_FAMILY_MEMBERS ids, so a cap-2 finalist set of
- * two large duplicate families could otherwise fan out to ~80 `in:comments`
- * SEARCH queries per iteration (secondary-rate-limited, not the ~1s REST call the
- * cost ceiling assumed) with no bound but MAX_FAMILY_MEMBERS × cap ×
- * MAX_REVERSE_ITERATIONS. Bounding it here — counting distinct probes across the
- * whole fixpoint via the shared `alreadyProbed` set — makes the documented gh
- * ceiling a real bound. Overflow probe ids are treated as NOT-probed (no edge,
- * no blocker), which fails toward MORE candidates — the same safe direction the
- * handled-id budget takes — with a stderr note and a run-record line.
- */
-export const MAX_REVERSE_PROBE_QUERIES = 40;
-
-/**
- * Per-search page size for the reverse `in:comments` probes. 5x headroom over the
- * live queue scale, the same class of fix as #1808: a `--limit` sized to today's
- * ledger silently drops page-2 hits, so a terminal sibling that falls past the
- * first page is missed and its candidate burns another autofix run. The bound
- * stays finite by design — we do NOT paginate-until-found. Instead a search that
- * comes back with a FULL page (`hits.length >= REVERSE_SEARCH_LIMIT`, meaning
- * there may be a page 2 we never read) flips the function's `truncated` flag, so
- * the existing run-record surface reports the shortfall exactly as the probe
- * budget does — bounded-and-surfaced, matching #1808.
- */
-export const REVERSE_SEARCH_LIMIT = 100;
-
-/**
- * How many DISTINCT stub reads (`gh issue view` via `readVerdictCached`) the whole
- * reverse fixpoint may spend admitting hits. MAX_REVERSE_PROBE_QUERIES bounds the
- * `in:comments` SEARCHES, but each search returns up to REVERSE_SEARCH_LIMIT (100)
- * rows and every unseen row costs one authoritative verdict re-read before the
- * fence can reject it — so without this the verify-read leg fans out to
- * MAX_REVERSE_PROBE_QUERIES × REVERSE_SEARCH_LIMIT (40 × 100 = 4000) subprocesses,
- * and the pipeline note's "~40 cached verify reads" ceiling term is a hope, not a
- * bound. Bounding it here — counting only cache-MISS reads across the fixpoint via
- * the shared `verifyBudget`, so a re-encountered stub stays free — makes the
- * documented gh ceiling real for this leg too. At the ceiling the remaining hits
- * are treated as NOT-admitted (no edge, no blocker), which fails toward MORE
- * candidates — the same safe direction the probe and handled-id budgets take — and
- * flips the same `truncated` flag so the shortfall surfaces on the run record.
- */
-export const MAX_REVERSE_VERIFY_READS = 40;
-
-/**
- * Reverse-verify the FINALISTS' families (PR #1810 bug B). The forward
- * `duplicate_of` graph misses two shapes the collapse must still catch: a
- * finalist that declares NOTHING but is named by a handled sibling, and a hub id
- * two stubs share through an issue that is not a candidate. For each probe id
- * (the finalists' family member ids), ONE search
- * `"<ID>" in:comments label:"sentry-triage"` surfaces the stubs whose comments
- * reference it.
- *
- * A hit is admitted ONLY after the authoritative recheck: its verdict is
- * re-parsed through the same fence, and the probed id must actually appear in
- * that parsed `duplicate_of` (via `declaredFamilyIds` — the same project-scoped,
- * self-excluding, MAX_DUPLICATE_LOOKUPS-bounded set the forward path consumes),
- * with a title that parses to a valid SHORT-ID — a casual comment mention can
- * never forge an edge. An admitted hit joins the hub's WHOLE declared family
- * into the collapse's union (an edge from the hit to each of its local declared
- * ids, not just the probed one — so a hub H that names both the finalist P and a
- * terminal sibling Q pulls Q into P's family, where the caller's handled-recheck
- * can then read Q's own marker and stand the family down; probing P alone left
- * Q's edge on the floor). An admitted hit that ALSO carries a terminal marker
- * becomes a BLOCKER (its key joins handledShortIds) directly.
- *
- * Fail-SOFT, same direction as `openAutofixPrExists`: a `gh` failure on one
- * probe skips that probe this run (at worst one self-terminating extra attempt),
- * never rejecting out of selection. `alreadyProbed` carries across the caller's
- * fixpoint iterations so no id is queried twice AND so the per-run
- * MAX_REVERSE_PROBE_QUERIES budget bounds the whole fixpoint, not one call;
- * `verifyBudget` likewise carries across iterations so the per-run
- * MAX_REVERSE_VERIFY_READS cap bounds the hit-verification reads across the whole
- * fixpoint, not one probe. Returns
- * `{ edges: [[hitKey, declaredKey]], blockers: [hitKey], truncated }` —
- * `truncated` true when the per-run probe budget was reached (some finalists left
- * unverified), OR a single probe came back with a full REVERSE_SEARCH_LIMIT page
- * (a sibling may sit on an unread page 2), OR the per-run verify-read budget was
- * exhausted (some hits left unread and so un-admitted). All three leave the
- * fixpoint incomplete, failing toward MORE candidates, and all surface on the
- * same run-record line.
- */
-export async function reverseVerifyFamilies(
-  runGh,
-  repo,
-  probeIds,
-  options = {},
-) {
-  const project = options.project;
-  const stubCache = options.stubCache ?? new Map();
-  const probed = options.alreadyProbed ?? new Set();
-  // Shared across the caller's fixpoint iterations (like `probed`), so the
-  // per-run verify-read cap bounds the whole reverse leg, not one call. Absent
-  // (a standalone call), it defaults fresh, preserving cap-at-N per call.
-  const verifyBudget = options.verifyBudget ?? {
-    remaining: MAX_REVERSE_VERIFY_READS,
-  };
-  const edges = [];
-  const blockers = new Set();
-  let truncated = false;
-  for (const rawId of Array.isArray(probeIds) ? probeIds : []) {
-    const probeId = familyKey(rawId);
-    // Foreign-project, bare-slug, or shape-invalid ids never probe: they cannot
-    // be a local family member, and the key is also interpolated into a stderr
-    // note below, where a newline would inject a workflow command.
-    if (!isLocalFamilyId(probeId, project)) continue;
-    if (probed.has(probeId)) continue;
-    // Per-run probe budget (counts distinct LOCAL probes across the fixpoint via
-    // the shared `probed` set). At the ceiling, stop probing and treat the rest
-    // as not-probed — fewer blockers, MORE candidates, the safe direction.
-    if (probed.size >= MAX_REVERSE_PROBE_QUERIES) {
-      truncated = true;
-      break;
-    }
-    probed.add(probeId);
-    let hits;
-    try {
-      const stdout = await runGh([
-        "issue",
-        "list",
-        "--repo",
-        repo,
-        "--state",
-        "all",
-        "--search",
-        `"${probeId}" in:comments label:"${SENTRY_TRIAGE_QUEUE_LABEL}"`,
-        "--json",
-        "number,title,labels",
-        "--limit",
-        String(REVERSE_SEARCH_LIMIT),
-      ]);
-      hits = JSON.parse(stdout);
-    } catch (err) {
-      const message = err instanceof Error ? err.message : String(err);
-      process.stderr.write(
-        `note: reverse family search for ${probeId} failed (${message}); skipping this probe.\n`,
-      );
-      continue;
-    }
-    // A full page means `--limit` capped what the API returned before any
-    // client-side filter ran: a page-2 hit — possibly the terminal sibling —
-    // went unread. Surface it (do NOT paginate); the finalist gets one bounded,
-    // self-terminating extra attempt rather than a silent miss (the #1808 class).
-    if (Array.isArray(hits) && hits.length >= REVERSE_SEARCH_LIMIT) {
-      truncated = true;
-    }
-    for (const hit of Array.isArray(hits) ? hits : []) {
-      const hitShortId = parseShortId(hit.title ?? "");
-      if (!isValidShortId(hitShortId)) continue;
-      const hitKey = familyKey(hitShortId);
-      if (hitKey === probeId) continue;
-      // Per-run verify-read budget: a hit already in the cache is free (no
-      // subprocess), but an UNSEEN one costs a `gh issue view`. At the cap, leave
-      // the remaining unseen hits unread — un-admitted, so fewer blockers and
-      // MORE candidates — and surface it on the run record. `stubCache.has` mirrors
-      // `readVerdictCached`'s own cache check exactly, so the budget is charged for
-      // precisely the calls that spawn a subprocess.
-      if (!stubCache.has(String(hit.number))) {
-        if (verifyBudget.remaining <= 0) {
-          truncated = true;
-          continue;
-        }
-        verifyBudget.remaining -= 1;
-      }
-      const parsed = await readVerdictCached(
-        runGh,
-        repo,
-        hit.number,
-        stubCache,
-      );
-      if (!parsed) continue;
-      // The hub's WHOLE declared local family (project-scoped, self-excluded,
-      // MAX_DUPLICATE_LOOKUPS-bounded — the forward path's exact rule). The probe
-      // is admitted only if the hub genuinely names it; then EVERY declared id
-      // joins the hub, so a terminal sibling the hub names alongside the finalist
-      // is pulled into the same family for the caller's handled-recheck.
-      const hubFamily = declaredFamilyIds(
-        hitShortId,
-        parsed.duplicateOf ?? [],
-        project,
-      );
-      if (!hubFamily.includes(probeId)) continue;
-      for (const declaredKey of hubFamily) edges.push([hitKey, declaredKey]);
-      const labels = (hit.labels ?? [])
-        .map((label) => (typeof label === "string" ? label : label?.name))
-        .filter(Boolean);
-      if (
-        labels.includes(FIX_PR_OPENED_LABEL) ||
-        labels.includes(FIX_REFUSED_LABEL)
-      ) {
-        blockers.add(hitKey);
-      }
-    }
-  }
-  if (truncated) {
-    process.stderr.write(
-      `note: reverse family verification was truncated this run — the per-run probe budget (${MAX_REVERSE_PROBE_QUERIES}) or verify-read budget (${MAX_REVERSE_VERIFY_READS}) was reached, or a probe returned a full page; the unreached hits/probes are treated as not-admitted (fails toward MORE candidates).\n`,
-    );
-  }
-  return { edges, blockers: [...blockers], truncated };
 }
 
 /** Read a queue stub's title/labels/comments so it can be evaluated in full. */
